+++ src/modules/Extensions/ProstoPleer.hpp
@@ -76,7 +76,7 @@ public:
 	QList<AddressPrefix> addressPrefixList(bool);
 	void convertAddress(const QString &prefix, const QString &url, const QString &param, QString *stream_url, QString *name, QImage *img, QString *extension, IOController<> *ioCtrl);
 
-	QAction *getAction(const QString &, int, const QString &, const QString &, const QString &);
+	QAction *getAction(const QString &, double, const QString &, const QString &, const QString &);
 private:
 	ProstoPleerW w;
 };
